Extract common graph action modal
